+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 #### Other changes
 - Cleaned up README.md to reflect latest changes.
+- Curried functions (functions curried via `curry*` functions) now retain their arity property value (`length` value).
 
 ### 1.3.0
 - Added `noop` (no-op (op as in operation)) method (useful as a placeholder for variables/properties that should always contain a function).